agentcore/requirements.txt
strands-agents>=1.6.0
kubernetes>=28.0.0
requests>=2.31.0
boto3>=1.34.0
bedrock-agentcore>=1.4.3
